requirements.txt
asgiref==3.7.2
astroid==2.13.5
autopep8==2.0.0
Brotli==1.1.0
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
dill==0.3.8
dj-database-url==2.1.0
Django==4.1.3
django-cors-headers==3.13.0
djangorestframework==3.14.0
gunicorn==21.2.0
h11==0.14.0
idna==3.6
isort==5.13.2
lazy-object-proxy==1.10.0
mccabe==0.7.0
packaging==24.0
platformdirs==4.2.0
psycopg2-binary==2.9.9
pycodestyle==2.11.1
pylint==2.15.5
pylint-django==2.5.3
pylint-plugin-utils==0.8.2
pytz==2024.1
requests==2.31.0
sqlparse==0.4.4
tomli==2.0.1
tomlkit==0.12.3
typing_extensions==4.10.0
tzdata==2024.1
urllib3==2.2.1
uvicorn==0.29.0
whitenoise==6.6.0
wrapt==1.16.0
